added forget password logic

+++ src/main/java/com/carwashbackend/carWashMajorProjectBackend/service/CleanerService.java
@@ -3,13 +3,14 @@ package com.carwashbackend.carWashMajorProjectBackend.service;
 
 import com.carwashbackend.carWashMajorProjectBackend.entity.*;
 import com.carwashbackend.carWashMajorProjectBackend.repository.CleanerJPARepository;
+//import com.carwashbackend.carWashMajorProjectBackend.repository.CleanerotpJPARepository;
+import com.carwashbackend.carWashMajorProjectBackend.repository.CleanerotpJPARepository;
 import com.carwashbackend.carWashMajorProjectBackend.repository.WashedCarMediaExteriorAndInteriorRepository;
 import com.carwashbackend.carWashMajorProjectBackend.repository.WashedCarMediaRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.dao.DataAccessException;
 import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.dao.EmptyResultDataAccessException;
-import org.springframework.data.crossstore.ChangeSetPersister;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.security.authentication.BadCredentialsException;
@@ -18,13 +19,14 @@ import org.springframework.stereotype.Service;
 import org.springframework.web.multipart.MultipartFile;
 import org.springframework.web.servlet.support.ServletUriComponentsBuilder;
 
-import javax.swing.text.html.Option;
 import java.io.ByteArrayInputStream;
 import java.io.IOException;
 import java.nio.file.Files;
 import java.nio.file.Path;
 import java.nio.file.Paths;
 import java.text.SimpleDateFormat;
+import java.time.Duration;
+import java.time.LocalDateTime;
 import java.util.*;
 
 @Service
@@ -46,7 +48,17 @@ public class CleanerService {
     private WashedCarMediaExteriorAndInteriorRepository washedCarMediaExteriorAndInteriorRepository;
 
 
+    @Autowired
+    private MailService mailService;
+
+//    @Autowired
+//    private CleanerotpJPARepository cleanerotpJPARepository;
+
+    @Autowired
+    private CleanerotpJPARepository cleanerotpJPARepository;
     public int cnt = 0;
+
+
     public ResponseEntity<String> addCleaner(Cleaner cleaner, MultipartFile imageData, MultipartFile adhaarData) {
 
         if(cleanerJPARepository.existsById(cleaner.getEmail())) {
@@ -70,7 +82,7 @@ public class CleanerService {
         try {
             List<Cleaner> cleaners = cleanerJPARepository.findAll();
             System.out.println("printing all cleaners");
-            System.out.println(cleaners);
+//            System.out.println(cleaners);
             return new ResponseEntity<List<Cleaner>>(cleaners, HttpStatus.OK);
         } catch (Exception e) {
             return new ResponseEntity<>(HttpStatus.INTERNAL_SERVER_ERROR);
@@ -81,8 +93,24 @@ public class CleanerService {
     public ResponseEntity<String> deleteByemail(String email) {
 
         try {
-            cleanerJPARepository.deleteById(email);
-            return new ResponseEntity<>("User deleted successfully", HttpStatus.OK);
+            Optional<Cleaner> cleaner = cleanerJPARepository.findById(email);
+            if(cleaner.isPresent()) {
+
+                List<Rating> allCleanerRatings = cleaner.get().getAllCleanerRatings();
+                List<Car> allCleanerCars = cleaner.get().getAllCleanerCars();
+
+                for(int i = 0; i < allCleanerRatings.size(); i++) {
+                    allCleanerRatings.get(i).setCleaner(null);
+                }
+                for(int i = 0; i < allCleanerCars.size(); i++) {
+                    allCleanerCars.get(i).setAssigned(false);
+                    allCleanerCars.get(i).setCleaner(null);
+                }
+
+                cleanerJPARepository.deleteById(email);
+                return new ResponseEntity<>("User deleted successfully", HttpStatus.OK);
+            }
+            return new ResponseEntity<>("User not found", HttpStatus.NOT_FOUND);
         } catch (EmptyResultDataAccessException e) {
             // Handle the case where the entity with the specified ID was not found
             return new ResponseEntity<>("Error: User not found", HttpStatus.NOT_FOUND);
@@ -308,7 +336,29 @@ public class CleanerService {
 
                 cleaner.get().setCurrAdd(newCurrentAddress);
                 cleaner.get().setPermanentAdd(newPermanentAddress);
-                cleanerJPARepository.save(cleaner.get());
+
+                try {
+
+                    cleanerJPARepository.save(cleaner.get());
+                    String from = "majorp1apl@gmail.com";
+                    String to = "jainansh2510@gmail.com";
+                    String subject = "Address changed by cleaner " + cleaner.get().getName();
+                    String message = "<html><body>" +
+                            "<h1 style='color: #5e9ca0;'>Hello Admin,</h1>" +
+                            "<p style='color: #5e9ca0;'>A Cleaner has recently changed his address.</p>" +
+                            "<p style='color: #5e9ca0;'>Here's Cleaner information:</p>" +
+                            "<p>" + "cleaner name -> " + cleaner.get().getName() + "</p>" +
+                            "<p>" + "cleaner phone -> " + cleaner.get().getPhone() + "</p>" +
+                            "<p style='color: #5e9ca0;'>If you think this was mistake kindly contact cleaner as soon as possible</p>" +
+                            "<p style='color: #5e9ca0;'>Thank you for your attention.</p>" +
+                            "<p style='color: #5e9ca0;'>Best regards,<br>Washify team</p>" +
+                            "</body></html>";
+                    mailService.send(from, to, message, subject);
+                } catch (RuntimeException e) {
+                    e.printStackTrace();
+                }
+
+
 
                 return new ResponseEntity<>(HttpStatus.OK);
             } catch (RuntimeException e) {
@@ -378,5 +428,144 @@ public class CleanerService {
         }
 
         return new ResponseEntity<>(HttpStatus.INTERNAL_SERVER_ERROR);
+    }
+
+    // forget password
+
+    public ResponseEntity<Void> isUserValid(String username) {
+//        System.out.println("Yes is uservalid " + username);
+
+        Optional<Cleaner> cleaner = cleanerJPARepository.findById(username);
+//        System.out.println("yes it is valid");
+
+        if(cleaner.isPresent()) {
+            sendOtp(username);
+            return new ResponseEntity<>(HttpStatus.FOUND);
+        }
+        return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+    }
+
+    public static String generateOtp() {
+        // Generate a random 5-digit OTP
+        Random random = new Random();
+        int otp = 10000 + random.nextInt(90000); // Generates a random number between 10000 and 99999
+        return String.valueOf(otp);
+    }
+
+    public static LocalDateTime calculateExpiryTime() {
+        // Calculate expiry time as current time + 5 minutes
+        return LocalDateTime.now().plusMinutes(5);
+    }
+
+    private String composeOtpMessage(String user, String otp) {
+        return "<!DOCTYPE html>"
+                + "<html lang=\"en\">"
+                + "<head>"
+                + "<meta charset=\"UTF-8\">"
+                + "<meta name=\"viewport\" content=\"width=device-width, initial-scale=1.0\">"
+                + "<title>OTP Email</title>"
+                + "<style>"
+                + "body { font-family: Arial, sans-serif; background-color: #f5f5f5; color: #333; }"
+                + ".container { max-width: 600px; margin: 0 auto; padding: 20px; background-color: #fff; border-radius: 5px; box-shadow: 0 0 10px rgba(0, 0, 0, 0.1); }"
+                + "h1 { color: #007bff; }"
+                + "p { font-size: 16px; line-height: 1.5; }"
+                + ".otp { font-size: 24px; font-weight: bold; color: #28a745; }"
+                + "</style>"
+                + "</head>"
+                + "<body>"
+                + "<div class=\"container\">"
+                + "<h1>Verify Your Account</h1>"
+                + "<p>Hello " + user + ",</p>"
+                + "<p>Thank you for using our service! To complete your verification process, please use the following OTP:</p>"
+                + "<p class=\"otp\">" + otp + "</p>"
+                + "<p>This OTP is valid for 5 minutes only. Please do not share it with anyone for security reasons.</p>"
+                + "<p>If you did not request this OTP, please ignore this message.</p>"
+                + "<p>Best regards,<br>Washify</p>"
+                + "</div>"
+                + "</body>"
+                + "</html>";
+    }
+
+    public void sendOtp(String username) {
+
+        System.out.println("uyes yaahan bhi valid h" + username);
+        Optional<Cleaner> cleaner = cleanerJPARepository.findById(username);
+
+        if(cleaner.isPresent()) {
+            String otp = generateOtp();
+            LocalDateTime expiryTime = calculateExpiryTime();
+
+            System.out.println(cleaner.get().getEmail());
+            String from = "majorp1apl@gmail.com";
+            String to = cleaner.get().getEmail();
+            String subject = "Otp for password change ";
+            String message = composeOtpMessage(cleaner.get().getEmail(), otp);
+            mailService.send(from, to, message, subject);
+
+            // store the otp
+            Optional<Cleanerotp> cleanerotp = cleanerotpJPARepository.findById(cleaner.get().getEmail());
+//            Cleanerotp cleanerotp = cleanerotpJPARepository.findBycleanerUsername(cleaner.get().getEmail());
+            if(!cleanerotp.isPresent()) {
+                Cleanerotp cleanerotp1 = new Cleanerotp();
+                cleanerotp1.setCleanerUsername(username);
+                cleanerotp1.setOtp(otp);
+                cleanerotp1.setExpiresAt(expiryTime);
+                cleanerotpJPARepository.save(cleanerotp1);
+
+            } else {
+                cleanerotp.get().setExpiresAt(expiryTime);
+                cleanerotp.get().setOtp(otp);
+                cleanerotpJPARepository.save(cleanerotp.get());
+            }
+//            cleanerotp.setOtp(otp);
+//            cleanerotp.setExpiredAt(expiryTime);
+        }
+    }
+
+    public ResponseEntity<String> isValidOtp(Map<String, String> data) {
+        String username = data.get("username");
+        Optional<Cleanerotp> cleanerotp = cleanerotpJPARepository.findById(username);
+        if(cleanerotp.isPresent()) {
+
+            String userOtp = data.get("otp");
+            if(userOtp.equals(cleanerotp.get().getOtp())) {
+
+                LocalDateTime nowTime = LocalDateTime.now();
+                Duration duration = Duration.between(cleanerotp.get().getExpiresAt(), nowTime);
+
+                if (duration.getSeconds() <= 5 * 60) {
+
+                    System.out.println("The difference is not more than 5 minutes.");
+
+                    return new ResponseEntity<>("OTP is valid", HttpStatus.OK);
+                } else {
+                    System.out.println("The difference is more than 5 minutes.");
+                    return new ResponseEntity<>("OTP is expired", HttpStatus.CONFLICT);
+                }
+            }
+
+
+            return new ResponseEntity<>("Wrong Otp", HttpStatus.CONFLICT);
+        }
+
+        return new ResponseEntity<>("No such user exists", HttpStatus.NOT_FOUND);
+    }
+
+    public ResponseEntity<String> changePasswordLogin(Map<String, String> data) {
+
+        String username = data.get("username");
+        Optional<Cleaner> cleaner = cleanerJPARepository.findById(username);
+        if(cleaner.isPresent()) {
+            try {
+
+                String newPassword = data.get("newPassword");
+                cleaner.get().setPassword(passwordEncoder.encode(newPassword));
+                cleanerJPARepository.save(cleaner.get());
+                return new ResponseEntity<>("Password changed succesfully", HttpStatus.OK);
+            } catch (RuntimeException e) {
+                e.printStackTrace();
+            }
+        }
+        return new ResponseEntity<>("User not present", HttpStatus.NOT_FOUND);
     }
 }
